requirements.txt
tqdm
PyYAML
easydict
numpy
opencv-python
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorboardX